requirements.txt
flask
requests
pandas
ta
python-dotenv
